+++ config/config.exs
@@ -95,7 +95,7 @@ config :ex_audit,
     WraftDoc.Documents.Block,
     WraftDoc.ContentTypes.ContentType,
     WraftDoc.DataTemplates.DataTemplate,
-    WraftDoc.Documents.Field,
+    WraftDoc.Fields.Field,
     WraftDoc.Documents.Instance,
     WraftDoc.Documents.Instance.Version,
     WraftDoc.Documents.OrganisationField,
