+++ lib/wraft_doc/content_types/content_type_field.ex
@@ -9,7 +9,7 @@ defmodule WraftDoc.ContentTypes.ContentTypeField do
 
   schema "content_type_field" do
     belongs_to(:content_type, WraftDoc.ContentTypes.ContentType)
-    belongs_to(:field, WraftDoc.Documents.Field)
+    belongs_to(:field, WraftDoc.Fields.Field)
 
     timestamps()
   end
