+++ lib/wraft_doc/documents/documents.ex
@@ -18,7 +18,6 @@ defmodule WraftDoc.Documents do
   alias WraftDoc.Documents.ContentCollaboration
   alias WraftDoc.Documents.Counter
   alias WraftDoc.Documents.Engine
-  alias WraftDoc.Documents.Field
   alias WraftDoc.Documents.FieldType
   alias WraftDoc.Documents.Instance
   alias WraftDoc.Documents.Instance.History
@@ -32,6 +31,7 @@ defmodule WraftDoc.Documents do
   alias WraftDoc.Enterprise.Flow.State
   alias WraftDoc.Enterprise.Organisation
   alias WraftDoc.Enterprise.StateUser
+  alias WraftDoc.Fields.Field
   alias WraftDoc.Layouts.Layout
   alias WraftDoc.ProsemirrorToMarkdown
   alias WraftDoc.Repo
@@ -41,46 +41,6 @@ defmodule WraftDoc.Documents do
   alias WraftDoc.Workers.EmailWorker
   alias WraftDocWeb.Mailer
   alias WraftDocWeb.Mailer.Email
-
-  @doc """
-  Creates a field.
-
-  ## Example
-
-  iex> create_field(%FieldType{}, %{name: "name"})
-  {:ok, %Field{}}
-
-  iex> create_field(%FieldType{}, %{})
-  {:error, %Ecto.Changeset{}}
-  """
-  def create_field(field_type, params) do
-    field_type
-    |> build_assoc(:fields)
-    |> Field.changeset(params)
-    |> Repo.insert()
-  end
-
-  # TODO write test
-  @doc """
-    Update a field
-  """
-  @spec update_field(Field.t(), map) :: Field.t() | nil
-  def update_field(%Field{} = field, params) do
-    field
-    |> Field.update_changeset(params)
-    |> Repo.update()
-  end
-
-  # TODO write test
-  @doc """
-    Get field
-  """
-  @spec get_field(Ecto.UUID.t()) :: Field.t() | nil
-  def get_field(<<_::288>> = field_id) do
-    Repo.get(Field, field_id)
-  end
-
-  def get_field(_), do: nil
 
   @doc """
   List all engines.
