+++ lib/wraft_doc/documents/field_type.ex
@@ -13,7 +13,7 @@ defmodule WraftDoc.Documents.FieldType do
     embeds_many(:validations, WraftDoc.Validations.Validation, on_replace: :delete)
     belongs_to(:creator, WraftDoc.Account.User)
 
-    has_many(:fields, WraftDoc.Documents.Field)
+    has_many(:fields, WraftDoc.Fields.Field)
     has_many(:organisation_fields, WraftDoc.Documents.OrganisationField)
     timestamps()
   end
